Benchmark.rails helper
