+++ rails/benchmarks/activerecord_create.rb
@@ -1,21 +1,12 @@
 require_relative 'support/activerecord_base.rb'
+require_relative 'support/benchmark_rails.rb'
 
-m = Benchmark.measure do
+Benchmark.rails(1000, "activerecord/#{db_adapter}/create") do
   notes = ActiveRecord::Faker::LOREM.join ' '
-  10.times do
-    fields = {
-      name: ActiveRecord::Faker.name,
-      notes: notes,
-      created_at: Date.today
-    }
-    Exhibit.create(fields)
-  end
+  fields = {
+    name: ActiveRecord::Faker.name,
+    notes: notes,
+    created_at: Date.today
+  }
+  Exhibit.create(fields)
 end
-
-stats = {
-  component: "activerecord/#{db_adapter}/create",
-  version: ActiveRecord::VERSION::STRING,
-  timing: m.real
-}
-
-puts stats.to_json
