+++ rails/benchmarks/activerecord_fetch.rb
@@ -1,16 +1,11 @@
 require_relative 'support/activerecord_base.rb'
+require_relative 'support/benchmark_rails.rb'
 
-m = Benchmark.measure do
+m = Benchmark.rails(100, "activerecord/#{db_adapter}/fetch") do
   Exhibit.look Exhibit.limit(100)
   Exhibit.look Exhibit.first(100)
   Exhibit.feel Exhibit.limit(100).includes(:user)
   Exhibit.look Exhibit.limit(10000)
 end
 
-stats = {
-  component: "activerecord/#{db_adapter}/fetch",
-  version: ActiveRecord::VERSION::STRING,
-  timing: m.real
-}
-
-puts stats.to_json
+puts m.to_json
