+++ rails/benchmarks/activerecord_finders.rb
@@ -1,18 +1,13 @@
 require_relative 'support/activerecord_base.rb'
+require_relative 'support/benchmark_rails.rb'
 
 User.create!(name: 'kir', email: 'shatrov@me.com')
 
-m = Benchmark.measure do
+m = Benchmark.rails(100, "activerecord/#{db_adapter}/finders") do
   User.first
   User.find(1)
   User.find_by(email: 'shatrov@me.com')
   User.find_by(name: 'kir')
 end
 
-stats = {
-  component: "activerecord/#{db_adapter}/finders",
-  version: ActiveRecord::VERSION::STRING,
-  timing: m.real
-}
-
-puts stats.to_json
+puts m.to_json
