+++ rails/benchmarks/activerecord_new.rb
@@ -1,16 +1,11 @@
 require_relative 'support/activerecord_base.rb'
+require_relative 'support/benchmark_rails.rb'
 
-m = Benchmark.measure do
+m = Benchmark.rails(100, "activerecord/#{db_adapter}/new") do
   attrs = { name: 'sam' }
 
   Exhibit.new
   Exhibit.new(attrs)
 end
 
-stats = {
-  component: "activerecord/#{db_adapter}/new",
-  version: ActiveRecord::VERSION::STRING,
-  timing: m.real
-}
-
-puts stats.to_json
+puts m.to_json
