+++ rails/benchmarks/activerecord_scope.rb
@@ -1,13 +1,8 @@
 require_relative 'support/activerecord_base.rb'
+require_relative 'support/benchmark_rails.rb'
 
-m = Benchmark.measure do
+m = Benchmark.rails(100, "activerecord/#{db_adapter}/scope") do
   Exhibit.limit(10).with_name.with_notes
 end
 
-stats = {
-  component: "activerecord/#{db_adapter}/scope",
-  version: ActiveRecord::VERSION::STRING,
-  timing: m.real
-}
-
-puts stats.to_json
+puts m.to_json
