+++ rails/benchmarks/url_generation_others.rb
@@ -1,17 +1,11 @@
 require_relative 'support/url_generation_base.rb'
+require_relative 'support/benchmark_rails.rb'
 
-m = Benchmark.measure do
+m = Benchmark.rails(100, "actionpack/url_generation/others") do
   router = Router.new
 
   router.listing_redirect_path(:any)
   router.residential_path(id: 2, addressing_slug: 'brooklyn')
   router.by_category_professionals_path(2)
 end
-
-stats = {
-  component: 'actionpack/url_generation',
-  version: Rails.version.to_s,
-  timing: m.real
-}
-
-puts stats.to_json
+puts m.to_json
