+++ rails/benchmarks/url_generation_url_for.rb
@@ -1,17 +1,12 @@
 require_relative 'support/url_generation_base.rb'
+require_relative 'support/benchmark_rails.rb'
 
-m = Benchmark.measure do
+m = Benchmark.rails(100, "actionpack/url_generation/url_for") do
   router = Router.new
-  
+
   router.url_for(controller: 'prices', action: 'info')
   router.url_for(controller: 'coupons', action: 'some')
   router.url_for(controller: 'professionals', action: 'index')
 end
 
-stats = {
-  component: 'actionpack/url_generation',
-  version: Rails.version.to_s,
-  timing: m.real
-}
-
-puts stats.to_json
+puts m.to_json
